all: separate detections files
