+++ util.go
@@ -2,16 +2,11 @@ package box
 
 import (
 	"fmt"
-	"io/ioutil"
-	"log"
-	"os"
-	"runtime"
 
 	"strings"
 
 	"github.com/gookit/color"
 	"github.com/mattn/go-runewidth"
-	"github.com/xo/terminfo"
 )
 
 // expandedLine stores a tab-expanded line, and its visible length.
@@ -82,113 +77,6 @@ func repeatWithString(c string, n int, str string) string {
 	bar := strings.Repeat(c, count)
 	strNew := fmt.Sprintf(" %s %s", str, bar)
 	return strNew
-}
-
-// errorMsg prints the msg to os.Stderr and uses Red ANSI Color too if supported
-func errorMsg(msg string) {
-	// If the terminal doesn't supports the basic 4 bit
-	if detectTerminalColor() == terminfo.ColorLevelNone || (buildNumber < 10586 || winVersion < 10) {
-		fmt.Fprintln(os.Stderr, msg)
-	} else {
-		fmt.Fprintln(os.Stderr, color.Red.Sprint(msg))
-	}
-}
-
-func detectTerminalColor() terminfo.ColorLevel {
-	// Detect WSL as it has True Color support
-	wsl, err := ioutil.ReadFile("/proc/sys/kernel/osrelease")
-	if err != nil {
-		log.Fatal(err)
-	}
-	// Microsoft for WSL and microsoft for WSL 2
-	if strings.Contains(string(wsl), "microsoft") && strings.Contains(string(wsl), "Microsoft") {
-		return terminfo.ColorLevelMillions
-	}
-	level, err := terminfo.ColorLevelFromEnv()
-	if err != nil {
-		log.Fatal(err)
-	}
-	return level
-}
-
-// roundOffColorVertical rounds off the 24 bit Color to the terminals maximum color capacity for Vertical.
-func (b Box) roundOffColorVertical(col color.RGBColor) string {
-	if runtime.GOOS != "windows" {
-		// Check if the terminal supports 256 Colors only
-		if detectTerminalColor() == terminfo.ColorLevelHundreds {
-			return col.C256().Sprint(b.Vertical)
-			// Check if the terminal supports 16 Colors only
-		} else if detectTerminalColor() == terminfo.ColorLevelBasic {
-			return col.C16().Sprint(b.Vertical)
-			// Check if the terminal supports True Color
-		} else if detectTerminalColor() == terminfo.ColorLevelMillions {
-			return col.Sprint(b.Vertical)
-		} else {
-			// Return with a warning as the terminal doesn't supports color effect
-			errorMsg("[warning]: terminal does not support colors, using no effect")
-			return b.Vertical
-		}
-	} else {
-		// Before Windows Build Number 10586, console never supported ANSI Colors
-		if buildNumber < 10586 || winVersion < 10 {
-			errorMsg("[warning]: terminal does not support colors, using no effect")
-			return b.Vertical
-		} else {
-			if buildNumber >= 14931 {
-				// True Color is only possible after Windows 10 Build Number 14931
-				// Virtual Terminal Processing is also enabled
-				return col.Sprint(b.Vertical)
-
-			} else {
-				// After Windows 10 Build Number 10586 and if not upgraded to at least 14931 then round off
-				// True Color to 8 bit Color
-				return col.C256().Sprint(b.Vertical)
-			}
-		}
-	}
-}
-
-// roundOffColorVertical rounds off the 24 bit Color to the terminals maximum color capacity for TopBar and BottomBar.
-func roundOffColor(col color.RGBColor, topBar, bottomBar string) (string, string) {
-	if runtime.GOOS != "windows" {
-		// Check if the terminal supports 256 Colors only
-		if detectTerminalColor() == terminfo.ColorLevelHundreds {
-			TopBar := col.C256().Sprint(topBar)
-			BottomBar := col.C256().Sprint(bottomBar)
-			return TopBar, BottomBar
-			// Check if the terminal supports 16 Colors only
-		} else if detectTerminalColor() == terminfo.ColorLevelBasic {
-			TopBar := col.C16().Sprint(topBar)
-			BottomBar := col.C16().Sprint(bottomBar)
-			return TopBar, BottomBar
-			// Check if the terminal supports True Color
-		} else if detectTerminalColor() == terminfo.ColorLevelMillions {
-			TopBar := col.Sprint(topBar)
-			BottomBar := col.Sprint(bottomBar)
-			return TopBar, BottomBar
-		} else {
-			// Return with a warning as the terminal supports no Color
-			errorMsg("[warning]: terminal does not support colors, using no effect")
-			return topBar, bottomBar
-		}
-	} else {
-		if buildNumber < 10586 || winVersion < 10 {
-			// Before Build Number 10586, console never supported ANSI Colors
-			fmt.Fprintln(os.Stderr, "[warning]: terminal does not support colors, using no effect")
-			return topBar, bottomBar
-		} else {
-			if buildNumber >= 14931 {
-				// True Color is only possible after Windows 10 Build 14931
-				// Virtual Terminal Processing is enabled by default in the later versions
-				return col.Sprint(topBar), col.Sprint(bottomBar)
-
-			} else {
-				// After Windows 10 build 10586 and if not upgraded to at least then round off
-				// True Color to 8 bit Color
-				return col.C256().Sprint(topBar), col.C256().Sprint(bottomBar)
-			}
-		}
-	}
 }
 
 // checkColorType checks the type of b.Color then from the preferences and options
